+++ lib/gitingest.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
-require_relative "getingest/version"
+require_relative "gitingest/version"
+require_relative "gitingest/generator"
 
-module Getingest
+module Gitingest
   class Error < StandardError; end
   # Your code goes here...
 end